sdo client refactor
